tehtävät 4.15-23

+++ osa4/controllers/blogs.js
@@ -1,28 +1,40 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { tokenExtractor, userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: 'no title or url' });
   }
 
+
+  if (!user) {
+    return response.status(400).json({ error: 'no users found' })
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
@@ -37,17 +49,28 @@ blogsRouter.put('/:id', async (request, response) => {
   if (!updatedBlog) {
     return response.status(404).json({ error: 'blog not found' })
   }
-    response.json(updatedBlog);
+  response.json(updatedBlog);
 });
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blogToDelete = await Blog.findByIdAndDelete(request.params.id)
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  const user = request.user
+  if (!user) {
+    return response.status(400).json({ error: 'token missing or invalid' })
+  }
 
   if (!blogToDelete) {
     return response.status(404).json({ error: 'blog not found' })
   }
-  
+
+  if (blogToDelete.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'only the creator can delete this blog' })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 });
